Customer_service/requirements.txt
Django==4.2.6                # Django framework
djangorestframework==3.14.0   # Django REST Framework for APIs
psycopg2-binary==2.9.6       # PostgreSQL Database Adapter (Only if using PostgreSQL)
kafka-python==2.0.2          # Kafka for event-driven messaging
redis==4.5.1                 # Redis for caching
celery==5.2.7                # Celery for background tasks
django-cors-headers==4.2.0   # CORS support for frontend integration
gunicorn==21.2.0             # Production WSGI server
Pillow==10.0.0             # Image processing library